backend/requirements.txt
Django
djangorestframework
psycopg2-binary
python-dotenv
requests
django-cors-headers 